+++ src/Http/Parser/LumenRouteParams.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Tymon\JWTAuth\Http;
+namespace Tymon\JWTAuth\Http\Parser;
 
 use Illuminate\Support\Arr;
 use Illuminate\Http\Request;
