+++ src/Http/Parser/QueryString.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Tymon\JWTAuth\Http;
+namespace Tymon\JWTAuth\Http\Parser;
 
 use Illuminate\Http\Request;
 use Tymon\JWTAuth\Contracts\Http\Parser as ParserContract;
