+++ tests/Claims/FactoryTest.php
@@ -14,6 +14,7 @@ namespace Tymon\JWTAuth\Test\Validators;
 use Tymon\JWTAuth\Claims\Custom;
 use Tymon\JWTAuth\Claims\Factory;
 use Tymon\JWTAuth\Claims\Subject;
+use Tymon\JWTAuth\Claims\Expiration;
 use Tymon\JWTAuth\Test\AbstractTestCase;
 
 class FactoryTest extends AbstractTestCase
@@ -34,6 +35,7 @@ class FactoryTest extends AbstractTestCase
     public function it_should_get_a_defined_claim_instance_when_passing_a_name_and_value()
     {
         $this->assertInstanceOf(Subject::class, $this->factory->get('sub', 1));
+        $this->assertInstanceOf(Expiration::class, $this->factory->get('exp', $this->testNowTimestamp + 3600));
     }
 
     /** @test */
